pyproject.toml
[tool.poetry]
name = "can-explorer"
version = "0.1.9"
description = "Visualize CAN bus payloads in real time"
license = "GPL-3.0-or-later"
authors = ["TJ <tbruno25@gmail.com>"]
# maintainers = ["TJ <tbruno25@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Tbruno25/can-explorer"
keywords = ["can", "bus", "canbus", "can_bus", "dbc"]

[tool.poetry.dependencies]
python = ">=3.8, <4.0"
pyserial = "^3.5"
python-can = "^4.1.0"
dearpygui = "^1.9.0"
dearpygui-ext = "^0.9.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
mypy = "^1.2.0"
pillow = "^10.3.0"
pyautogui = "^0.9.54"
opencv-python = "^4.8.0.74"
pyvirtualdisplay = [
    {version = "^3.0", platform = "linux"},
    {version = "^3.0", platform = "darwin"},
]
msvc-runtime = {version = "^14.34.31931", platform = "windows"}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.poetry.scripts]
can-explorer = "can_explorer.__main__:__main__"
